mycelix-workspace/mycelix-core/papers/reproducibility/requirements_exact.txt
# Mycelix Paper Reproducibility - Exact Dependencies
# These versions were used to generate all paper results
# Last verified: January 8, 2026

# Core scientific computing
numpy==1.26.3
scipy==1.11.4

# Data handling
pandas==2.1.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.1

# Machine learning (for FL simulation)
scikit-learn==1.3.2

# Configuration
pyyaml==6.0.1

# Statistics
statsmodels==0.14.1

# Progress bars
tqdm==4.66.1

# Type checking (optional, for development)
typing-extensions==4.9.0

# Async support
asyncio-throttle==1.0.2

# JSON handling
orjson==3.9.10

# Hash computation
xxhash==3.4.1

# Table generation
tabulate==0.9.0

# LaTeX table generation
pylatex==1.4.2

# Note: PyTorch is optional for reproducibility
# Paper benchmarks use simulated gradients by default
# Uncomment below for full model training experiments:
# torch==2.1.2
# torchvision==0.16.2
